+++ frontend/src/app/(auth)/sign-in/[[...sign-in]]/page.tsx
