+++ frontend/src/app/(auth)/sign-up/[[...sign-up]]/page.tsx
