+++ frontend/src/app/(protected)/hello/page.tsx
@@ -1,10 +1,13 @@
 "use client";
 import React from "react";
 import { useAuth } from "@clerk/nextjs";
+import { useCourses, useCreateCourse } from "@/hooks/useCourses";
 
 const Page = () => {
   const { isLoaded, userId } = useAuth();
+  const { data: courses, isLoading, error } = useCourses(userId || "");
 
+  const createCourse = useCreateCourse(); // can call createCourse.mutate()
   if (!isLoaded) {
     return <div>Loading...</div>;
   }
@@ -14,6 +17,7 @@ const Page = () => {
       <p className="font-mono">This is a protected route</p>
       {/* <p>{JSON.stringify(user)}</p> */}
       <p>{JSON.stringify(userId)}</p>
+      <p>{JSON.stringify(courses)}</p>
     </>
   );
 };
